requirements.txt
google-generativeai
torch
transformers
Pillow
gtts
opencv-python
deep-translator
